# common-platform-rule-engine-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: common-platform-rule-engine
  description: Rule Engine Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/rules
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/constants/{constantId}:
    get:
      tags:
        - Constants
      summary: Get constant by ID
      description: Retrieve a system constant by its ID
      operationId: getConstantById
      parameters:
        - name: constantId
          in: path
          description: Constant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Constant retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConstantDTO'
        '404':
          description: Constant not found
        '500':
          description: Internal server error
    put:
      tags:
        - Constants
      summary: Update a constant
      description: Update an existing system constant
      operationId: updateConstant
      parameters:
        - name: constantId
          in: path
          description: Constant ID
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConstantDTO'
        required: true
      responses:
        '200':
          description: Constant updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConstantDTO'
        '400':
          description: Invalid constant data
        '404':
          description: Constant not found
        '500':
          description: Internal server error
    delete:
      tags:
        - Constants
      summary: Delete a constant
      description: Delete a system constant by ID
      operationId: deleteConstant
      parameters:
        - name: constantId
          in: path
          description: Constant ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Constant deleted successfully
        '404':
          description: Constant not found
        '500':
          description: Internal server error
...
  /api/v1/constants:
    post:
      tags:
        - Constants
      summary: Create a new constant
      description: Create a new system constant with UPPER_CASE_WITH_UNDERSCORES naming convention. Constants are automatically loaded in YAML DSL rules when referenced.
      operationId: createConstant
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConstantDTO'
        required: true
      responses:
        '201':
          description: Constant created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConstantDTO'
        '400':
          description: Invalid constant data or naming convention violation
        '409':
          description: Constant with the same code already exists
        '500':
          description: Internal server error
  /api/v1/constants/filter:
    post:
      tags:
        - Constants
      summary: Filter constants
      description: Retrieve constants based on filtering criteria
      operationId: filterConstants
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestConstantDTO'
        required: true
      responses:
        '200':
          description: Constants retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
        '500':
          description: Internal server error
...
  /api/v1/constants/code/{code}:
    get:
      tags:
        - Constants
      summary: Get constant by code
      description: Retrieve a system constant by its unique code
      operationId: getConstantByCode
      parameters:
        - name: code
          in: path
          description: Constant code
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Constant retrieved successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ConstantDTO'
        '404':
          description: Constant not found
        '500':
          description: Internal server error
...
components:
  schemas:
...
    ConstantDTO:
      required:
        - code
        - name
        - required
        - valueType
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the constant
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        code:
          maxLength: 100
          minLength: 2
          pattern: ^[A-Z][A-Z0-9_]*$
          type: string
          description: Unique constant code following UPPER_CASE_WITH_UNDERSCORES convention. Used in YAML DSL rules to reference this constant.
          example: MIN_CREDIT_SCORE
        name:
          maxLength: 255
          minLength: 2
          type: string
          description: Human-readable name for the constant
          example: Minimum Credit Score
        valueType:
          type: string
          description: Data type of the constant value
          example: NUMBER
          enum:
            - STRING
            - NUMBER
            - BOOLEAN
            - DATE
            - OBJECT
        required:
          type: boolean
          description: Whether this constant must have a value for rule evaluation
          example: true
        description:
          maxLength: 1000
          minLength: 0
          type: string
          description: Optional description explaining the purpose of this constant
          example: Minimum credit score required for loan approval
        currentValue:
          type: object
          description: Current value of the constant. Type should match the valueType field.
          example: 650
        createdAt:
          type: string
          description: Timestamp when the constant was created
          format: date-time
          example: '2025-01-12T10:30:00Z'
        updatedAt:
          type: string
          description: Timestamp when the constant was last updated
          format: date-time
          example: '2025-01-12T15:45:00Z'
      description: System constant used in rule evaluation. Constants follow UPPER_CASE_WITH_UNDERSCORES naming convention.
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestConstantDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/ConstantDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...